Audio: use double for seeking

+++ src/audio/aldatasource.h
@@ -44,7 +44,7 @@ struct ALDataSource
 
 	/* If the source doesn't support seeking, it will
 	 * reset back to the beginning */
-	virtual void seekToOffset(float seconds) = 0;
+	virtual void seekToOffset(double seconds) = 0;
 
 	/* The frame count right after wrap around */
 	virtual uint32_t loopStartFrames() = 0;
